LayoutTests/platform/glib/imported/w3c/web-platform-tests/content-security-policy/reporting/report-frame-ancestors-no-parent-cookies.sub-expected.txt



PASS Violation report status OK.
FAIL Test report cookies. assert_equals: Report should not contain any cookies expected (string) "None" but got (object) object "[object Object]"
